Add #1404 When asset is requested from web via bkclientjs, it should appear in new search tab - init open search bar and add a new tab in handle_bkclientjs_get_asset()
